+++ ForceOps.Lib/src/ForceOpsContext/ILoggerFactory.cs
@@ -1,0 +1,8 @@
+using Serilog;
+
+namespace ForceOps.Lib;
+
+public interface ILoggerFactory
+{
+	public ILogger CreateLogger<T>();
+}
