+++ ForceOps.Lib/src/ForceOpsContext/IProcessKiller.cs
