+++ ForceOps.Lib/src/ForceOpsContext/LoggerFactory.cs
@@ -1,0 +1,15 @@
+﻿using Serilog;
+
+namespace ForceOps.Lib;
+
+public class LoggerFactory : ILoggerFactory
+{
+	readonly ILogger logger = new LoggerConfiguration()
+		   .WriteTo.Console()
+		   .CreateLogger();
+
+	public ILogger CreateLogger<T>()
+	{
+		return logger;
+	}
+}
