+++ ForceOps.Lib/src/ForceOpsContext/RelaunchAsElevated.cs
